requirements.txt
fastapi
uvicorn
python-multipart
pandas
pytest
httpx
pydantic
google-generativeai
python-dotenv
